pyproject.toml
[project]
name = "reasoning"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "biomcp-python>=0.1.10",
    "elevenlabs>=2.7.1",
    "fastapi>=0.116.1",
    "future>=1.0.0",
    "graspologic>=3.3.0",
    "ipykernel>=6.30.1",
    "jsonschema>=4.23.0",
    "llama-cloud-services>=0.6.50",
    "llama-index-embeddings-azure-inference>=0.3.0",
    "llama-index-embeddings-bedrock>=0.5.2",
    "llama-index-embeddings-cohere>=0.5.1",
    "llama-index-embeddings-gemini>=0.3.2",
    "llama-index-embeddings-huggingface-api>=0.3.1",
    "llama-index-embeddings-openai>=0.3.1",
    "llama-index-graph-stores-neo4j>=0.4.6",
    "llama-index-indices-managed-llama-cloud>=0.6.11",
    "llama-index-llms-openai>=0.4.7",
    "llama-index-observability-otel>=0.1.1",
    "llama-index-tools-mcp>=0.2.6",
    "llama-index-workflows>=1.1.0",
    "loguru>=0.7.3",
    "mcp[cli]>=1.6.0",
    "mrkdwn-analysis>=0.1.6",
    "neo4j>=5.28.2",
    "numpy>=1.21.6,<1.28.0",
    "openai>=1.76.0",
    "openai-agents>=0.0.17",
    "opentelemetry-exporter-otlp-proto-http>=1.35.0",
    "psycopg2-binary>=2.9.10",
    "pydub>=0.25.1",
    "pytest-asyncio>=0.24.0",
    "python-dotenv>=1.1.0",
    "tenacity>=8.5.0",
    "pyvis>=0.3.2",
    "qdrant-client<1.15.0",
    "randomname>=0.2.1",
    "scipy>=1.9.0,<1.12.0",
    "streamlit>=1.44.1",
    "textual>=5.3.0",
    "watchdog>=6.0.0",
]

pythonpath = ["."]
